requirements.txt
selenium==4.33.0
pytest
allure-pytest
webdriver-manager
requests